+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Additional strategy to set viewport size [Trello 1919](https://trello.com/c/9vRolTeu)
 - Option for use latest available versions of libraries for Py3 [Trello 2008](https://trello.com/c/mt8wSiQC)
+- Visual Viewport support for UFG [Trello 1957](https://trello.com/c/jWvdBwex)
 ### Updated
 - Improved traceback info for Configuration.add_browser/s() and match regions [Trello 1998](https://trello.com/c/8IoYsgNI)